+++ src/biv_lite/__init__.py
@@ -1,2 +1,4 @@
 from .biv_mesh import BivMesh, Components
 from .biv_frames import BivFrames
+from .biv_parametric import BivParametric
+from .meshing.biv_motion_ui import BivMotionUI