+++ src/biv_lite/biv_frames.py
@@ -3,26 +3,6 @@ from pathlib import Path
 from typing import List
 import re
 from collections.abc import Sequence
-import numpy as np
-from scipy.interpolate import make_splrep
-
-
-def perinterp(x: np.array, y: np.array, k: int = 1):
-    """Periodic 1D interpolation.
-
-    :param x: x samples as 1D array
-    :param y: y samples as 1D array
-    :param k: number of samples to be replicated beginning & end
-
-    :return: a function
-    """
-
-    # append the first k-sample to the end and last k-sample to the beginning
-    xs = np.concat((x[-k:] - 1.0, x, 1.0 + x[:k]))
-    ys = np.concat((y[-k:], y, y[:k]))
-
-    # return the interpolation function
-    return make_splrep(xs, ys)
 
 
 class BivFrames(Sequence):
@@ -31,7 +11,7 @@ class BivFrames(Sequence):
     """
     def __init__(self, biv_mesh_list: List[BivMesh], frames = None):
         self.biv_mesh = biv_mesh_list
-        if not frames:
+        if frames is None:
             self.frames = list(range(len(self.biv_mesh)))
         else:
             self.frames = frames