+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 * Created `BivFrames` class which is a list of `BivMesh` with some extra props & methods.
 * Added UI app
 * Added interpolation function for BivFrames
+* Created `BivParametric` class which can generate interpolated biventricular models in continuous temporal domain t=[0,1].
 
 ### Refactoring
 * `biv_plots.py` -> `plots.py`
@@ -23,3 +24,4 @@
 * Plot a complete biv mesh is now written in `vis.py` file
 * `to_pyvista_faces` function is inside `utils.py` file
 * All typer app's are in `app` subfolder, except for `main.py`
+* `BivMotionUI` class is separated in `biv_motion_ui.py` file